refactor(player-options): rename rows/ to panes/
Pure rename: each file under rows/ becomes the same file under panes/, plus the mod declaration in mod.rs is updated. The submodule names (build_main_rows, build_advanced_rows, build_uncommon_rows) keep their existing function names so callers are untouched.
