+++ src/screens/player_options/panes/advanced.rs
