+++ src/screens/player_options/panes/uncommon.rs
